+++ src/app/sitemap.ts
@@ -70,6 +70,7 @@ const pages = [
   { path: '/resources/public-safety-software-finland', priority: 0.75 },
   { path: '/resources/public-safety-software-portugal', priority: 0.75 },
   { path: '/resources/public-safety-software-turkey', priority: 0.75 },
+  { path: '/resources/public-safety-software-austria', priority: 0.75 },
   { path: '/resources/public-safety-software-uae', priority: 0.75 },
   { path: '/resources/public-safety-software-sweden', priority: 0.75 },
   { path: '/resources/public-safety-software-south-korea', priority: 0.75 },